requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.27
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
pydantic>=2.5.0
pydantic-settings>=2.2.0
jinja2>=3.1.3
aiofiles>=23.2.1
dlib>=19.24.0
face-recognition>=1.3.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

